move function to database_loader module
