+++ alphaquant/resources/database_loader.py
@@ -68,6 +68,7 @@ def get_swissprot_path( organism = "human"):
 
 def _get_path_to_database( database_name, organism):
     database_folder = os.path.join(pathlib.Path(__file__).parent.absolute(), "reference_databases")
+    LOGGER.info(f"Checking for reference databases in {database_folder}")
     if not os.path.exists(database_folder):
         LOGGER.info(f"Downloading reference databases to {database_folder}")
         try:
@@ -80,3 +81,23 @@ def _get_path_to_database( database_name, organism):
     if not os.path.exists(database_path):
         raise Exception(f"Reference database {database_name} for organism {organism} not found at {database_path}")
     return database_path
+
+
+def load_dl_predicted_phosphoprone_sequences(organism = "human"):
+    organism_map = {"human": "human_uniprot_reviewed_phos_prob.tsv"}
+    database_folder = os.path.join(pathlib.Path(__file__).parent.absolute(), "..","resources","phosphopred_databases")
+
+    LOGGER.info(f"Checking for phosphopred databases in {database_folder}")
+    if not os.path.exists(database_folder):
+        LOGGER.info(f"Downloading phosphopred databases to {database_folder}")
+        try:
+            dsd = ab_downloader.DataShareDownloader("https://datashare.biochem.mpg.de/s/stH9pmNe6O9CRHG", output_dir=f"{database_folder}/..")
+            dsd.download()
+        except Exception as e:
+            LOGGER.error(f"Failed to download phosphopred databases: {str(e)}")
+            raise Exception(f"Failed to download phosphopred databases: {str(e)}") from e
+    database_path = os.path.join(database_folder, organism_map[organism])
+
+    df_phospho_predlib = pd.read_csv(database_path, sep='\t')
+    df_phospho_predlib["sequence"] = [f"SEQ_{x}_" for x in df_phospho_predlib["sequence"]]
+    return set(df_phospho_predlib[df_phospho_predlib['ptm_prob'] > 0.5]["sequence"])
